Refatorando depósitos e saques para operation
Run transaction não é necessário em operation mas na transação
